# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: CarbonPulse API — Precision carbon intelligence platform
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users/me:
    get:
      operationId: getMe
      tags: [users]
      summary: Get current user profile
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          description: Not found
    post:
      operationId: upsertMe
      tags: [users]
      summary: Create or update user profile (called on first login)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserInput"
      responses:
        "200":
          description: User upserted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    patch:
      operationId: updateMe
      tags: [users]
      summary: Update user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdate"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"

  # ── ACTIVITIES ────────────────────────────────────────────────────────────────
...
  /insights:
    get:
      operationId: listInsights
      tags: [insights]
      summary: List AI insights (most recent first)
      parameters:
        - name: limit
          in: query
          schema: { type: integer, default: 8 }
      responses:
        "200":
          description: List of insights
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Insight"
  /insights/latest:
    get:
      operationId: getLatestInsight
      tags: [insights]
      summary: Get the most recent AI insight
      responses:
        "200":
          description: Latest insight
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Insight"
        "404":
          description: No insight yet
...
components:
  schemas:
...
    User:
      type: object
      required: [id, email, region, weeklyBudgetKg, createdAt]
      properties:
        id:            { type: string }
        email:         { type: string }
        displayName:   { type: ["string", "null"] }
        avatarUrl:     { type: ["string", "null"] }
        region:        { type: string }
        weeklyBudgetKg: { type: number }
        createdAt:     { type: string }
    UserInput:
      type: object
      required: [email]
      properties:
        email:       { type: string }
        displayName: { type: string }
        avatarUrl:   { type: string }
        region:      { type: string }
    UserUpdate:
      type: object
      properties:
        displayName:    { type: string }
        avatarUrl:      { type: string }
        region:         { type: string }
        weeklyBudgetKg: { type: number }

    # ── ACTIVITY ───────────────────────────────────────────────────────────────
...
    Insight:
      type: object
      required: [id, userId, weekStartDate, insightText, primaryDriver, suggestedAction, projectedSavingKgYear, totalCo2KgForWeek, deltaFromPrevWeek, createdAt]
      properties:
        id:                   { type: string }
        userId:               { type: string }
        weekStartDate:        { type: string }
        insightText:          { type: string }
        primaryDriver:        { type: string }
        suggestedAction:      { type: string }
        projectedSavingKgYear: { type: number }
        totalCo2KgForWeek:    { type: number }
        deltaFromPrevWeek:    { type: number }
        createdAt:            { type: string }

    # ── GOAL ───────────────────────────────────────────────────────────────────
...